I ACCIDENTALLY CLICKED COMMIT INSTEAD OF COMMIT AND PUSH. SO I WENT INTO TEAM AND PUSHED UP STREAM BUT I DON'T THINK MY COMMIT MESSAGE WAS SAVED SO HERE IT IS AGAIN
Currently working on the part I know is going to be the hardest:
creating and returning characters. As of right now, the program should
return a 2D array and display that 2D array in the JavaFX class.
Instead, I'm currently running into NullPointerException (more detail
can be found in the test spreadsheet). In addition to this, I spent the
class updating my Gantt chart and added the first test (I added in the
buttons and the list view so I was just testing the make sure the
program runs and the output is looking correct). Taking a break for now
but I will be back to continue working on this soon.